# specs/002-improve-wellmio-booking/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Wellmio Booking API
  version: 1.0.0
paths:
  /api/timeslots:
    get:
      summary: Get available timeslots for a given date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: A list of available timeslots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlot'
  /api/bookings:
    post:
      summary: Create a new booking and a Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                timeslot_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: The Stripe checkout session URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: url
...
  /api/stripe/webhook:
    post:
      summary: Handle Stripe webhook events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
components:
  schemas:
    TimeSlot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        is_booked:
          type: boolean
...