# src/app/api/timeslots/route.ts
import { NextResponse } from 'next/server';
import { createClient } from '@supabase/supabase-js';
import { TimeSlot } from '@/lib/db/schema';
import { Redis } from '@upstash/redis';

// Initialize Supabase client
const supabase = createClient(
  process.env.NEXT_PUBLIC_SUPABASE_URL || 'https://mock.supabase.co',
  process.env.SUPABASE_SERVICE_ROLE_KEY || 'mock-key'
);

/**
 * GET /api/timeslots
 *
 * Returns available time slots for booking.
 * Can be filtered by date.
 * Based on OpenAPI spec and contract tests.
 */
export async function GET(request: Request) {
  try {
    const { searchParams } = new URL(request.url);
    const date = searchParams.get('date');

    const cacheEnabled =
      process.env.NODE_ENV === 'production' &&
      !!process.env.UPSTASH_REDIS_REST_URL &&
      !!process.env.UPSTASH_REDIS_REST_TOKEN;
    const isTest =
      process.env.NODE_ENV === 'test' ||
      process.env.JEST_WORKER_ID !== undefined ||
      process.env.TEST_BASE_URL !== undefined;
    const cacheKey = `timeslots:available:v1:${date || 'all'}`;
    type ApiTimeSlot = Pick<TimeSlot, 'id' | 'start_time' | 'end_time'>;
    let cached: ApiTimeSlot[] | null = null;
    let redis: Redis | null = null;

    if (cacheEnabled && !isTest) {
      redis = new Redis({
        url: process.env.UPSTASH_REDIS_REST_URL!,
        token: process.env.UPSTASH_REDIS_REST_TOKEN!,
      });
      try {
        cached = await redis.get<ApiTimeSlot[]>(cacheKey);
        if (cached && Array.isArray(cached)) {
          return NextResponse.json(cached, {
            status: 200,
            headers: { 'Content-Type': 'application/json' },
          });
        }
      } catch {
        console.log('Redis unavailable, proceeding without cache');
      }
    }
    
    let query = supabase
      .from('timeslots')
      .select('*')
      .eq('is_booked', false)
      .gte('start_time', new Date().toISOString()) // Only future time slots
      .order('start_time', { ascending: true });

    if (date) {
        const startDate = new Date(date);
        const endDate = new Date(date);
        endDate.setDate(startDate.getDate() + 1);
        query = query.gte('start_time', startDate.toISOString()).lt('start_time', endDate.toISOString());
    }

    const { data, error } = await query;

    if (error) {
        console.log('Database error, using mock data:', error.message);
        // Return mock data for testing
        const mockTimeSlots: TimeSlot[] = [
            {
              id: '123e4567-e89b-12d3-a456-426614174000',
              start_time: new Date(Date.now() + 24 * 60 * 60 * 1000).toISOString(), // Tomorrow
              end_time: new Date(
                Date.now() + 24 * 60 * 60 * 1000 + 30 * 60 * 1000
              ).toISOString(), // Tomorrow + 30 min
              is_booked: false,
            },
            {
              id: '123e4567-e89b-12d3-a456-426614174001',
              start_time: new Date(
                Date.now() + 2 * 24 * 60 * 60 * 1000
              ).toISOString(), // Day after tomorrow
              end_time: new Date(
                Date.now() + 2 * 24 * 60 * 60 * 1000 + 30 * 60 * 1000
              ).toISOString(), // Day after tomorrow + 30 min
              is_booked: false,
            },
        ];
        const response: ApiTimeSlot[] = mockTimeSlots.map(slot => ({
            id: slot.id,
            start_time: slot.start_time,
            end_time: slot.end_time,
        }));
        return NextResponse.json(response, {
            status: 200,
            headers: {
              'Content-Type': 'application/json',
            },
        });
    }

    const response: ApiTimeSlot[] =
      data?.map(slot => ({
        id: slot.id,
        start_time: slot.start_time,
        end_time: slot.end_time,
      })) || [];

    if (redis) {
      try {
        await redis.set(cacheKey, response, { ex: 30 });
      } catch {}
    }

    return NextResponse.json(response, {
      status: 200,
      headers: {
        'Content-Type': 'application/json',
      },
    });
  } catch (error) {
    console.error('Unexpected error:', error);
    return NextResponse.json(
      { error: 'Internal server error' },
      { status: 500 }
    );
  }
}